make 'crates/runtime-datafusion/src/analyzer_rule/partitioned_table_scan_rewrite/'
